pyproject.toml
[project]
name = "assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "eshafayyazz", email = "eshafayyaz76@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.7",
    "openai-agents>=0.2.5",
]

[project.scripts]
assistant-agent = "assistant_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
